+++ app/models/project.rb
@@ -1,8 +1,9 @@
 class Project < ApplicationRecord
-  has_many :addons, class_name: 'ProjectAddon'
+  has_many :project_addons
   has_many :project_stencils
-  has_many :project_workshops
+  has_many :workshop_projects
+  has_many :addons, through: :project_addons
   has_many :stencils, through: :project_stencils
-  has_many :workshops, through: :project_workshops
+  has_many :workshops, through: :workshop_projects
   has_many_attached :project_images
 end