+++ app/models/workshop_project.rb
@@ -1,0 +1,4 @@
+class WorkshopProject < ApplicationRecord
+  belongs_to :project
+  belongs_to :workshop
+end